requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.30.0
alembic==1.13.1

# Pydantic & Validation
pydantic==2.10.6
pydantic-settings==2.1.0
email-validator==2.2.0

# AI & Transcription
anthropic==0.18.1
openai==1.10.0

# Audio Processing
ffmpeg-python==0.2.0
pydub==0.25.1
audioop-lts==0.2.1  # Required for pydub on Python 3.13+

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
stripe==8.0.0
apscheduler==3.10.4

# Document Generation
python-docx==1.1.0
reportlab==4.0.9

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.13
sendgrid==6.11.0
